requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==2.0
django-webpack-loader==0.5.0
djangorestframework==3.7.3
flake8==3.5.0
idna==2.6
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2017.3
requests==2.18.4
urllib3==1.22
